Tidy up a loose end in the header retirement plan.
If a server might be 30x-downgrading clients that don't send the header, then
the client has to keep sending the header on every request, even if it has
HSTS set.

Fortunately, it's easy for the server to tell the client that it won't do
that.  This request proposes doing so with a new HSTS directive.