Framed navigational requests ought to be upgraded.
As written, we weren't upgrading cross-origin framed requests, as they were treated
in the same way that we treated top-level framed requests. The carveout for navigations
shouldn't apply to framed navigations, as those are directly relevant to the embedder's
security state, and the directive ought to apply to them.

This patch closes w3c/webappsec-upgrade-insecure-requests#4.